+++ crate/kinematic_model/src/loaded.rs
@@ -1,0 +1,9 @@
+//! Types that represent processed configuration.
+
+pub use self::{
+    object_acceleration_sequence::{ObjectAccelerationSequence, ObjectAccelerationSequenceHandle},
+    object_acceleration_sequence_handles::ObjectAccelerationSequenceHandles,
+};
+
+mod object_acceleration_sequence;
+mod object_acceleration_sequence_handles;