+++ crate/object_model/src/loaded/object.rs
@@ -1,6 +1,7 @@
 use collision_model::loaded::{BodySequenceHandles, InteractionsSequenceHandles};
 use derivative::Derivative;
 use derive_new::new;
+use kinematic_model::loaded::ObjectAccelerationSequenceHandles;
 use sequence_model::loaded::{SequenceEndTransitions, WaitSequenceHandles};
 use spawn_model::loaded::SpawnsSequenceHandles;
 use sprite_model::loaded::SpriteRenderSequenceHandles;
@@ -11,6 +12,8 @@ use sprite_model::loaded::SpriteRenderSequenceHandles;
 pub struct Object {
     /// Handles to `WaitSequence`s that this object uses.
     pub wait_sequence_handles: WaitSequenceHandles,
+    /// Handles to `ObjectAccelerationSequence`s that this object uses.
+    pub object_acceleration_sequence_handles: ObjectAccelerationSequenceHandles,
     /// Handles to `SpriteRenderSequence`s that this object uses.
     pub sprite_render_sequence_handles: SpriteRenderSequenceHandles,
     /// Handles to `BodySequence`s that this object uses.