+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 ### Added
 
 * Support arbitrary sequence names. ([#155])
+* Support acceleration in configuration. ([#125])
 
 ### Changed
 
 * ***Breaking:*** Configuration files now use YAML instead of TOML. ([#154])
 
+[#125]: https://gitlab.com/azriel91/autexousious/issues/125
 [#154]: https://gitlab.com/azriel91/autexousious/issues/154
 [#155]: https://gitlab.com/azriel91/autexousious/issues/155